add the point process representation
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/pot@459 80a141b9-e137-0410-8b51-c7aaed2cc172

+++ pkg/src/header.h
@@ -11,7 +11,7 @@
 void gpdlik(double *data, int *n, double *loc, double *scale,
 	    double *shape, double *dns);
 void pplik(double *data, int *n, double *loc, double *scale,
-	   double *shape, double *thresh, int *noy, double *dns);
+	   double *shape, double *thresh, double *noy, double *dns);
 void samlmu(double *x, int *nmom, int *n, double *lmom);
 
 //From clust.c
